+++ src/main/java/prgms/jpamission2/domain/CustomerRepository.java
@@ -1,4 +1,4 @@
-package prgms.jpamission2.config.domain;
+package prgms.jpamission2.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
