+++ src/main/java/prgms/jpamission2/domain/Food.java
@@ -1,4 +1,4 @@
-package prgms.jpamission2.config.domain;
+package prgms.jpamission2.domain;
 
 import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Entity;
@@ -8,7 +8,7 @@ import lombok.Setter;
 @Setter
 @Getter
 @Entity
-@DiscriminatorValue("CAR")
-public class Car extends Item {
-    private int power;
+@DiscriminatorValue("FOOD")
+public class Food extends Item {
+    private String chef;
 }
