+++ src/main/java/prgms/jpamission2/domain/OrderStatus.java
@@ -1,4 +1,4 @@
-package prgms.jpamission2.config.domain;
+package prgms.jpamission2.domain;
 
 public enum OrderStatus {
     OPENED, CANCELLED
